Define character traits for sf::Uint8
Character traits are only standardized for character types of which
sf::Uint8 is not. All major C++ implementations happen to define
this specialization but because it is not standard C++ they are
allowed to remove it as LLVM has done by deprecating this specialization
in LLVM 18. It is removed in LLVM 19.